analyzer: move name tracking to types; road towards serializing types.
